GPIO class now working

+++ usr/gpio.h
@@ -10,7 +10,7 @@ typedef struct cfg {
 class gpio_base_class
 {
 public:
-    virtual int Init(GPIO_TypeDef *GPIO, uint16_t GPIO_PIN, uint16_t GPIO_MODE, uint16_t GPIO_SPEED) = 0;
+    virtual int Init(GPIO_TypeDef *GPIO, uint32_t GPIO_PIN, uint32_t GPIO_MODE, uint32_t GPIO_SPEED) = 0;
     gpio_cfg_t *cfg;
 };
 
@@ -18,7 +18,7 @@ class gpio_class : public gpio_base_class, public rcc_periph_class
 {
 public:
     gpio_class() { };
-    int Init(GPIO_TypeDef *GPIO, uint16_t GPIO_PIN, uint16_t GPIO_MODE, uint16_t GPIO_SPEED);
+    int Init(GPIO_TypeDef *GPIO, uint32_t GPIO_PIN, uint32_t GPIO_MODE, uint32_t GPIO_SPEED);
     int Init(void);
     int ClockEnable(void);
 };
